oval: create reference index for use with Lookup methods
This commit implements an index for the various Lookup methods. There
was a memoized lookup in previous versions of the Lookup methods, but
through an oversight it didn't make it into the committed version.

The runtime to construct the index is paid on the first call of a Lookup
method. A sync.Once is used to make sure it's only done once and that
the method remains concurrent-OK.

This also defers parsing the reference until after a lookup has failed,
which saves allocations and provides the same error messages.

This also fixes a broken test to use the correct, generic Lookup method.